--- src/test/java/StepDefinitions/TES22.java
@@ -1,81 +1,0 @@
-package StepDefinitions;
-
-import Pages.DialogContent;
-import Pages.LeftNav;
-import io.cucumber.datatable.DataTable;
-import io.cucumber.java.Scenario;
-import io.cucumber.java.en.And;
-import io.cucumber.java.en.Then;
-
-import java.awt.*;
-import java.util.List;
-
-public class TES22 {
-    DialogContent dc = new DialogContent();
-    LeftNav ln = new LeftNav();
-
-    @And("Click on the element in the left Nav")
-    public void clickOnTheElementInTheLeftNav(DataTable elements) {
-        List<String> listElement = elements.asList(String.class);
-
-        for (int i = 0; i < listElement.size(); i++) {
-            ln.findAndClick(listElement.get(i));
-
-        }
-    }
-
-    @And("Click on the element in the Dialog")
-    public void clickOnTheElementInTheDialog(DataTable elements) {
-
-        List<String> listElement = elements.asList(String.class);
-
-        for (int i = 0; i < listElement.size(); i++) {
-            dc.findAndClick(listElement.get(i));
-        }
-    }
-
-    @And("User sending the keys in Dialog content")
-    public void userSendingTheKeysInDialogContent(DataTable elements) {
-        List<List<String>> listElement = elements.asLists(String.class);
-
-        for (int i = 0; i < listElement.size(); i++)
-            dc.findAndSend(listElement.get(i).get(0), listElement.get(i).get(1));
-    }
-
-    @And("User sending the keys and edit in Dialog content")
-    public void userSendingTheKeysAndEditInDialogContent(DataTable elements) throws AWTException {
-        List<List<String>> listElement = elements.asLists(String.class);
-
-        for (int i = 0; i < listElement.size(); i++) {
-            dc.SearchAndClear(listElement.get(i).get(1));
-        }
-
-
-    }
-
-
-    @Then("Success message should be displayed")
-    public void successMessageShouldBeDisplayed() {
-        dc.findAndContainsText("successMessage", "success");
-    }
-
- 
-
-    @And("User edit and save in Dialog content")
-    public void userEditAndSaveInDialogContent(DataTable elements) {
-        List<List<String>> listElement = elements.asLists(String.class);
-
-        for (int i = 0; i < listElement.size(); i++) {
-            dc.findAndSend(listElement.get(i).get(0), listElement.get(i).get(1));
-        }
-    }
-    @And("User delete item from Dialog")
-    public void userDeleteItemFromDialog(DataTable elements) throws AWTException {
-        List<List<String>> listElement = elements.asLists(String.class);
-
-        for(int i=0;i<listElement.size();i++) {
-            dc.SearchAndDeletee(listElement.get(i).get(1));
-        }
-    }
-    }
-
